requirements.txt
alembic==1.15.2
annotated-types==0.7.0
blinker==1.9.0
click==8.1.8
feedparser==6.0.11
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
flask-openapi3==4.1.0
flask-openapi3-rapidoc==9.3.8
flask-openapi3-redoc==2.4.0
flask-openapi3-swagger==5.20.7
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
python-dotenv==1.1.0
sgmllib3k==1.0.0
SQLAlchemy==2.0.40
SQLAlchemy-Utils==0.41.2
typing-inspection==0.4.0
typing_extensions==4.13.0
Werkzeug==3.1.3
